added a new file
I added a new file called Home. which now has the count down clock and the button which will be linked to our profiles. 
I also removed the buttons from the index.html page and may have messed up the CSS for the (.login) but I'm not sure. Ill take a closer look soon.