new folder-new function
I added anew folder which will hold pictures, I also added a fade function which which make backgrown dark and will overlay a profile when you click on our names at the top. of page after logging in.

I will be moving that out of the site to add a better suted funtion which will display our resumes as PDF files